// frontend/src/api/model/presetMenuDto.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * YaoYao Dinner API
 * Restaurant management API
 * OpenAPI spec version: 1.0
 */
import type { PresetMenuItemDto } from './presetMenuItemDto';

export interface PresetMenuDto {
  id: string;
  price: number;
  isActive: boolean;
  items: PresetMenuItemDto[];
}