// frontend/src/api/model/presetMenuItemDto.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * YaoYao Dinner API
 * Restaurant management API
 * OpenAPI spec version: 1.0
 */

export interface PresetMenuItemDto {
  variantId: string;
  foodId: string;
  foodName: string;
  variantLabel: string;
  /** @nullable */
  price: number | null;
  currency: string;
  quantity: number;
}